refactor: Move metrics to a different module, rename function
